pyproject.toml
[project]
name = "rtv-ml-system"
version = "1.0.0"
description = "Multi-Agent AI System for Raising the Village - Text-to-SQL, RAG, and Orchestration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Jonah Kyagaba"},
]

dependencies = [
    # LLM & Agent Framework
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",

    # Database
    "duckdb>=1.1.0",

    # Vector Store & Embeddings
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",

    # RAG
    "langchain-text-splitters>=0.3.0",
    "python-docx>=1.1.0",
    "pypdf>=5.0.0",
    "unstructured>=0.15.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Caching & Memory
    "redis>=5.2.0",

    # Data Processing
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "numpy>=1.26.0",

    # Evaluation
    "ragas>=0.2.0",

    # Observability
    "structlog>=24.4.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",

    # Resilience
    "tenacity>=9.0.0",
    "pybreaker>=1.2.0",

    # SQL Validation
    "sqlglot>=25.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]
analysis = [
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "scipy>=1.14.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
